requirements.txt
asgiref==3.8.1
azure-core==1.31.0
azure-storage-blob==12.23.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
dj-database-url==2.2.0
Django==5.1.1
django-cors-headers==4.4.0
django-rest-authtoken==2.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
isodate==0.7.2
neo4j==5.25.0
neo4j-driver==5.25.0
pika==1.3.2
polyline==2.0.2
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
